+++ src/pkg/services/m365/api/events_pager.go
@@ -305,10 +305,11 @@ func (c Events) GetAddedAndRemovedItemIDs(
 	effectivePageSize := minEventsDeltaPageSize
 
 	logger.Ctx(ctx).Infow(
-		"retrying event item query with reduced page size",
+		"retrying list event item query with reduced page size",
 		"delta_pager_effective_page_size", effectivePageSize,
 		"delta_pager_default_page_size", c.options.DeltaPageSize)
 
+	// Get new pagers just to make sure we don't have partial state in them.
 	deltaPager = c.newEventsDeltaPagerWithPageSize(
 		ctx,
 		userID,
@@ -316,13 +317,49 @@ func (c Events) GetAddedAndRemovedItemIDs(
 		prevDeltaLink,
 		effectivePageSize,
 		idAnd()...)
+	pager = c.NewEventsPager(
+		userID,
+		containerID,
+		idAnd(lastModifiedDateTime)...)
+
+	addedRemoved, err = pagers.GetAddedAndRemovedItemIDs[models.Eventable](
+		ctx,
+		pager,
+		deltaPager,
+		prevDeltaLink,
+		config.CanMakeDeltaQueries,
+		config.LimitResults,
+		pagers.AddedAndRemovedByAddtlData[models.Eventable])
+	if err == nil || !errors.Is(err, graph.ErrServiceUnavailableEmptyResp) {
+		return addedRemoved, clues.Stack(err).OrNil()
+	}
+
+	logger.Ctx(ctx).Infow(
+		"retrying list event item query with non-delta pager",
+		"effective_page_size", maxNonDeltaPageSize)
+
+	// Attempt yet another fallback by using the regular pager if we still can't
+	// get data with a smaller page size. Get new pagers just to make sure we
+	// don't have partial state in them.
+	deltaPager = c.newEventsDeltaPagerWithPageSize(
+		ctx,
+		userID,
+		containerID,
+		prevDeltaLink,
+		effectivePageSize,
+		idAnd()...)
+	pager = c.NewEventsPager(
+		userID,
+		containerID,
+		idAnd(lastModifiedDateTime)...)
 
 	return pagers.GetAddedAndRemovedItemIDs[models.Eventable](
 		ctx,
 		pager,
 		deltaPager,
 		prevDeltaLink,
-		config.CanMakeDeltaQueries,
+		// Disable delta queries.
+		false,
 		config.LimitResults,
 		pagers.AddedAndRemovedByAddtlData[models.Eventable])
 }