+++ src/pkg/services/m365/api/graph/service_test.go
@@ -12,6 +12,7 @@ import (
 	"time"
 
 	"github.com/alcionai/clues"
+	"github.com/h2non/gock"
 	"github.com/microsoftgraph/msgraph-sdk-go/models"
 	"github.com/microsoftgraph/msgraph-sdk-go/users"
 	"github.com/stretchr/testify/assert"
@@ -25,6 +26,115 @@ import (
 	"github.com/alcionai/corso/src/pkg/errs/core"
 	graphTD "github.com/alcionai/corso/src/pkg/services/m365/api/graph/testdata"
 )
+
+// ---------------------------------------------------------------------------
+// Unit tests
+// ---------------------------------------------------------------------------
+
+type GraphUnitSuite struct {
+	tester.Suite
+}
+
+func TestGraphUnitSuite(t *testing.T) {
+	suite.Run(t, &GraphUnitSuite{
+		Suite: tester.NewUnitSuite(t),
+	})
+}
+
+func (suite *GraphUnitSuite) TestNoRetryPostNoContent404() {
+	const (
+		host    = "https://graph.microsoft.com"
+		retries = 3
+	)
+
+	t := suite.T()
+
+	ctx, flush := tester.NewContext(t)
+	t.Cleanup(flush)
+
+	a := tconfig.NewFakeM365Account(t)
+	creds, err := a.M365Config()
+	require.NoError(t, err, clues.ToCore(err))
+
+	// Run with a single retry since 503 retries are exponential and
+	// the test will take a long time to run.
+	service, err := NewGockService(
+		creds,
+		count.New(),
+		MaxRetries(1),
+		MaxConnectionRetries(retries))
+	require.NoError(t, err, clues.ToCore(err))
+
+	t.Cleanup(gock.Off)
+
+	gock.New(host).
+		Post("/v1.0/users").
+		Reply(http.StatusNotFound).
+		BodyString("").
+		Type("text/plain")
+
+	// Since we're retrying all 404s with no content the endpoint we use doesn't
+	// matter.
+	_, err = service.Client().Users().Post(ctx, models.NewUser(), nil)
+	assert.ErrorIs(t, err, ErrNotFoundEmptyResp)
+
+	assert.False(t, gock.IsPending(), "some requests not seen")
+}
+
+func (suite *GraphUnitSuite) TestRetryGetNoContent404() {
+	const (
+		host    = "https://graph.microsoft.com"
+		retries = 3
+
+		emptyUserList = `{
+"@odata.context": "https://graph.microsoft.com/v1.0/$metadata#users",
+"value": []
+}`
+	)
+
+	t := suite.T()
+
+	ctx, flush := tester.NewContext(t)
+	t.Cleanup(flush)
+
+	a := tconfig.NewFakeM365Account(t)
+	creds, err := a.M365Config()
+	require.NoError(t, err, clues.ToCore(err))
+
+	// Run with a single retry since 503 retries are exponential and
+	// the test will take a long time to run.
+	service, err := NewGockService(
+		creds,
+		count.New(),
+		MaxRetries(1),
+		MaxConnectionRetries(retries))
+	require.NoError(t, err, clues.ToCore(err))
+
+	t.Cleanup(gock.Off)
+
+	gock.New(host).
+		Get("/v1.0/users").
+		Times(retries - 1).
+		Reply(http.StatusNotFound).
+		BodyString("").
+		Type("text/plain")
+
+	gock.New(host).
+		Get("/v1.0/users").
+		Reply(http.StatusOK).
+		JSON(emptyUserList)
+
+	// Since we're retrying all 404s with no content the endpoint we use doesn't
+	// matter.
+	_, err = service.Client().Users().Get(ctx, nil)
+	assert.NoError(t, err, clues.ToCore(err))
+
+	assert.False(t, gock.IsPending(), "some requests not seen")
+}
+
+// ---------------------------------------------------------------------------
+// Integration tests
+// ---------------------------------------------------------------------------
 
 type GraphIntgSuite struct {
 	tester.Suite
@@ -93,7 +203,7 @@ func (suite *GraphIntgSuite) TestHTTPClient() {
 			checkConfig: func(t *testing.T, c *clientConfig) {
 				assert.Equal(t, defaultDelay, c.minDelay, "default delay")
 				assert.Equal(t, defaultMaxRetries, c.maxRetries, "max retries")
-				assert.Equal(t, defaultMaxRetries, c.maxConnectionRetries, "max connection retries")
+				assert.Equal(t, adapterMaxRetries, c.maxConnectionRetries, "max connection retries")
 			},
 		},
 		{
@@ -221,43 +331,92 @@ func (suite *GraphIntgSuite) TestAdapterWrap_catchesPanic() {
 	require.Contains(t, err.Error(), "panic", clues.ToCore(err))
 }
 
-func (suite *GraphIntgSuite) TestAdapterWrap_retriesConnectionClose() {
-	t := suite.T()
-
-	ctx, flush := tester.NewContext(t)
-	defer flush()
-
-	retryInc := 0
-
-	// the panics should get caught and returned as errors
-	alwaysECONNRESET := mwForceResp{
-		err: syscall.ECONNRESET,
-		alternate: func(req *http.Request) (bool, *http.Response, error) {
-			retryInc++
-			return false, nil, nil
+func (suite *GraphIntgSuite) TestAdapterWrap_retriesConnectionInterruptions() {
+	table := []struct {
+		name             string
+		providedErr      error
+		expectRetryCount int
+		expectErr        assert.ErrorAssertionFunc
+	}{
+		{
+			name:             "ECONNRESET",
+			providedErr:      syscall.ECONNRESET,
+			expectRetryCount: 7,
+			expectErr:        assert.Error,
+		},
+		{
+			name:             "connection reset by peer",
+			providedErr:      clues.New("connection reset by peer what is love"),
+			expectRetryCount: 7,
+			expectErr:        assert.Error,
+		},
+		{
+			name:             "read: connection timed out",
+			providedErr:      clues.New("read: connection timed out baby don't hurt me"),
+			expectRetryCount: 7,
+			expectErr:        assert.Error,
+		},
+		{
+			name:             "unexpected EOF",
+			providedErr:      io.ErrUnexpectedEOF,
+			expectRetryCount: 7,
+			expectErr:        assert.Error,
+		},
+		{
+			name:             "nil error",
+			providedErr:      nil,
+			expectRetryCount: 1,
+			expectErr:        assert.NoError,
+		},
+		{
+			name:             "non retriable error",
+			providedErr:      clues.New("no more"),
+			expectRetryCount: 1,
+			expectErr:        assert.Error,
 		},
 	}
 
-	adpt, err := CreateAdapter(
-		suite.credentials.AzureTenantID,
-		suite.credentials.AzureClientID,
-		suite.credentials.AzureClientSecret,
-		count.New(),
-		appendMiddleware(&alwaysECONNRESET),
-		// Configure retry middlewares so that they don't retry on connection reset.
-		// Those middlewares have their own tests to verify retries.
-		MaxRetries(-1))
-	require.NoError(t, err, clues.ToCore(err))
+	for _, test := range table {
+		suite.Run(test.name, func() {
+			t := suite.T()
 
-	// Retry delay doesn't really matter here since all requests will be intercepted
-	// by the test middleware. Set it to 0 to reduce test runtime.
-	aw := adpt.(*adapterWrap)
-	aw.retryDelay = 0
+			ctx, flush := tester.NewContext(t)
+			defer flush()
 
-	// the query doesn't matter
-	_, err = NewService(adpt).Client().Users().Get(ctx, nil)
-	require.ErrorIs(t, err, syscall.ECONNRESET, clues.ToCore(err))
-	require.Equal(t, 4, retryInc, "number of retries")
+			retryInc := 0
+			forceErrMW := mwForceResp{
+				err: test.providedErr,
+				// send some dummy response, because kiota retry handler panics
+				// if err is nil and resp is nil.
+				resp: &http.Response{},
+				alternate: func(req *http.Request) (bool, *http.Response, error) {
+					retryInc++
+					return false, nil, nil
+				},
+			}
+
+			adpt, err := CreateAdapter(
+				suite.credentials.AzureTenantID,
+				suite.credentials.AzureClientID,
+				suite.credentials.AzureClientSecret,
+				count.New(),
+				appendMiddleware(&forceErrMW),
+				// Configure retry middlewares so that they don't retry on connection reset.
+				// Those middlewares have their own tests to verify retries.
+				MaxRetries(-1))
+			require.NoError(t, err, clues.ToCore(err))
+
+			// Retry delay doesn't really matter here since all requests will be intercepted
+			// by the test middleware. Set it to 0 to reduce test runtime.
+			aw := adpt.(*adapterWrap)
+			aw.retryDelay = 0
+
+			// the query doesn't matter
+			_, err = NewService(adpt).Client().Users().Get(ctx, nil)
+			test.expectErr(t, err, clues.ToCore(err))
+			require.Equal(t, test.expectRetryCount, retryInc, "number of retries")
+		})
+	}
 }
 
 func (suite *GraphIntgSuite) TestAdapterWrap_retriesBadJWTToken() {
@@ -315,14 +474,14 @@ func (suite *GraphIntgSuite) TestAdapterWrap_retriesBadJWTToken() {
 		NewItemCalendarsItemEventsDeltaRequestBuilder("https://graph.microsoft.com/fnords/beaux/regard", adpt).
 		Get(ctx, nil)
 	assert.ErrorIs(t, err, core.ErrAuthTokenExpired, clues.ToCore(err))
-	assert.Equal(t, 4, retryInc, "number of retries")
+	assert.Equal(t, adapterMaxRetries+1, retryInc, "number of retries")
 
 	retryInc = 0
 
 	// the query doesn't matter
 	_, err = NewService(adpt).Client().Users().Get(ctx, nil)
 	assert.ErrorIs(t, err, core.ErrAuthTokenExpired, clues.ToCore(err))
-	assert.Equal(t, 4, retryInc, "number of retries")
+	assert.Equal(t, adapterMaxRetries+1, retryInc, "number of retries")
 }
 
 // TestAdapterWrap_retriesInvalidRequest tests adapter retries for graph 400
@@ -398,7 +557,7 @@ func (suite *GraphIntgSuite) TestAdapterWrap_retriesInvalidRequest() {
 				_, err = NewService(adpt).Client().Users().Get(ctx, nil)
 				return err
 			},
-			expectedRetries: 4,
+			expectedRetries: adapterMaxRetries + 1,
 		},
 		{
 			name: "POST request, no retry",